# Conflicts:
#	src/Entity/Sortie.php
#	src/Form/LieuType.php

+++ src/Controller/Sortiev2Controller.php
@@ -8,10 +8,11 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Security;
 
 class Sortiev2Controller extends AbstractController
 {
-    #[Route('/sortiev2/{id}', name: 'sortiev2_show')]
+    #[Route('/sortiev2/{id}', name: 'sortiev2_show',requirements: ['id'=> '\d+'] )]
     #[IsGranted("ROLE_USER")]
     public function show(int $id, SortieRepository $sortieRepository): Response
     {
@@ -31,14 +32,22 @@ class Sortiev2Controller extends AbstractController
 
     #[Route('/delete/{id}', name: 'sortie_delete',requirements: ['id'=> '\d+'])]
     #[IsGranted("ROLE_USER")]
-    public function delete(int $id, SortieRepository $sortieRepository)
+    public function delete(int $id, SortieRepository $sortieRepository, Security $security)
     {
+        $user=$security->getUser();
         $sortie = $sortieRepository->find($id);
+        if ($sortie->getUser()==$user){
+
+
 
         $sortieRepository->remove($sortie, true);
 
-        $this->addFlash('success', $sortie->getName()."has been removed !");
+        $this->addFlash('success', "la sortie à ".$sortie->getName()." a été supprimé !");
 
+        return $this->redirectToRoute('main_home');
+
+        }
+        $this->addFlash('error', "Vous n'êtes pas l'organisateur de cette sortie !");
         return $this->redirectToRoute('main_home');
 
     }
